+++ go-fuzz/sys_windows.go
@@ -4,10 +4,8 @@
 package main
 
 import (
-	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"reflect"
 	"syscall"
 	"unsafe"
@@ -46,13 +44,4 @@ func createMapping(name string, size int) (*Mapping, []byte) {
 func (m *Mapping) destroy() {
 	syscall.UnmapViewOfFile(m.addr)
 	syscall.CloseHandle(m.mapping)
-}
-
-func setupCommMapping(cmd *exec.Cmd, comm *Mapping, rOut, wIn *os.File) {
-	syscall.SetHandleInformation(syscall.Handle(comm.mapping), syscall.HANDLE_FLAG_INHERIT, 1)
-	syscall.SetHandleInformation(syscall.Handle(rOut.Fd()), syscall.HANDLE_FLAG_INHERIT, 1)
-	syscall.SetHandleInformation(syscall.Handle(wIn.Fd()), syscall.HANDLE_FLAG_INHERIT, 1)
-	cmd.Env = append(cmd.Env, fmt.Sprintf("GO_FUZZ_COMM_FD=%v", comm.mapping))
-	cmd.Env = append(cmd.Env, fmt.Sprintf("GO_FUZZ_IN_FD=%v", rOut.Fd()))
-	cmd.Env = append(cmd.Env, fmt.Sprintf("GO_FUZZ_OUT_FD=%v", wIn.Fd()))
 }
